assignment finished

+++ src/main/java/com/walt/model/DriverDistance.java
@@ -2,5 +2,6 @@ package com.walt.model;
 
 public interface DriverDistance {
     Driver getDriver();
-    Long getTotalDistance();
+//    Long getTotalDistance(); // error in definition of Type, in Delivery Object the distance is double
+    Double getTotalDistance();
 }
